Remove unused Plus feature views + strings
